int -> long

+++ datetime.cpp
@@ -100,7 +100,7 @@ void zero_pad (std::string &t)
 // YYYY-MM-DD hh:mm:ss
 // @param t1 start date-time string
 // @param t2 end date-time string
-int timediff (std::string t1, std::string t2)
+long int timediff (std::string t1, std::string t2)
 {
     int Y1 = atoi (t1.substr (0, 4).c_str ()),
         M1 = atoi (t1.substr (5, 2).c_str ()),
@@ -115,8 +115,8 @@ int timediff (std::string t1, std::string t2)
         m2 = atoi (t2.substr (14, 2).c_str ()),
         s2 = atoi (t2.substr (17, 2).c_str ());
 
-    int y1 = daynum (Y1, M1, D1), y2 = daynum (Y2, M2, D2);
-    int d1 = y1 * 3600 * 24 + h1 * 3600 + m1 * 60 + s1,
+    long int y1 = daynum (Y1, M1, D1), y2 = daynum (Y2, M2, D2);
+    long int d1 = y1 * 3600 * 24 + h1 * 3600 + m1 * 60 + s1,
         d2 = y2 * 3600 * 24 + h2 * 3600 + m2 * 60 + s2;
 
     return d2 - d1;
@@ -124,20 +124,22 @@ int timediff (std::string t1, std::string t2)
 
 // Julian day number calculation from
 // http://www.cs.utsa.edu/~cs1063/projects/Spring2011/Project1/jdn-explanation.html
-int daynum (int y, int m, int d)
+long int daynum (int y, int m, int d)
 {
     int a = static_cast <int> (floor ((14.0 - m) / 12.0));
     y = y + 4800 - a;
     m = m + 12 * a - 3;
-    int res = d + static_cast <int> (floor ((153 * m + 2) / 5)) + 365 * y +
-        static_cast <int> (floor (y / 4)) -
-        static_cast <int> (floor (y / 100)) +
-        static_cast <int> (floor (y / 400)) - 32045;
+    long int res = static_cast <long int> (d) +
+        static_cast <int> (floor ((153 * m + 2) / 5)) +
+        static_cast <long int> (365 * y) +
+        static_cast <long int> (floor (y / 4)) -
+        static_cast <long int> (floor (y / 100)) +
+        static_cast <long int> (floor (y / 400)) - 32045;
     return res;
 }
 
 // format output of timediff as days, hours, min, seconds
-std::string prettytime (int t)
+std::string prettytime (long int t)
 {
     std::string out;
     int days = static_cast <int> (floor (t / (3600 * 24)));
